+++ spec/lib/oauth/id_token_token_request_spec.rb
@@ -30,9 +30,7 @@ describe Doorkeeper::OAuth::IdTokenTokenRequest do
     pre_auth
   end
 
-  let(:owner) do
-    double :owner, id: 7866, to_i: 7866
-  end
+  let(:owner) { build_stubbed(:user) }
 
   # just to make sure self created pre_auth is authorizable
   it 'pre_auth should be valid' do