+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 - [#112] Add grant_types_supported to discovery response
 - [#111] Add configuration callback `select_account_for_resource_owner` to support the `prompt=select_account` param
 - [#117] Fix migration template to use Rails migrations DSL for association.
+- [#116] Support Doorkeeper API (> 5.4) for registering custom grant flows.
 
 ## v1.7.2 (2020-05-20)